Install GPG key from keyserver.ubuntu.com
deb.nodesource.com is now in CloudFront, and older versions of Python
(such as than installed in Ubuntu 12.04 and 14.04) can no longer install
the GPG key from that server. Instead, install the key from
keyserver.ubuntu.com, and include an ID so that downloading can be
skipped if the key is already installed. Fixes #33.